(broken) updating the editor
widget uses itself as the item editor.
the editor stack's template contains a navbar, but that's not ideal because we want to change the way the editor functions by switching from iconpicker to svgfilter, so we have to rebuild the widget, and replace the version in the editor stack with the new one.
doing that does not seem to be working. the PI shows a blank line where the array should be, and opening the editor stack after saving and closing shows a blank spot where the widget sits. the PI does not successfully render if the widget is selected, but it does render if the card or stack is selected.